backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
markdown==3.3.4
django-filter==2.4.0
django-cors-headers==3.7.0